### 2018-10-01 #254 @AldoCompagnoni
Submission: popler R package
Submitting Author: Aldo Compagnoni (@gAldoCompagnoni)
Repository:   https://github.com/AldoCompagnoni/popler/issues
Version submitted:   0.0.0.9001
Version accepted: TBD

---

`popler` provides a way to browse and download population and community time-series data associated with the `popler` online database, which is currently hosted on a public AWS server. This database formats about 300 time-series datasets from the USA Long-Term Ecological Research Netowork (LTER) using a common structure.

```
Package: popler
Title: Popler R Package
Version: 0.0.0.9001
  person("Compagnoni", "Aldo", email = "aldo.compagnoni@aggiemail.usu.edu",
  role = c("cre","aut")),
  person("Bibian", "Andrew", email = "ajbibian@rice.edu", role= c("aut")),
  person("Ochocki", "Brad", email = "brad.ochocki@rice.edu", role= c("aut")),
  person("Levin", "Sam", email = "sam.levin@idiv.de", role = c("aut")),
  person("Miller", "Tom", email = "tom.miller@rice.edu", role= c("aut")))
Description: Browse and query the popler database.
    dbplyr (>= 1.1.0),
    ggplot2 (>= 3.0.0),
    grid (>= 3.3.0),
    lazyeval (>= 0.2.0),
    lubridate (>= 1.7.4),
    magrittr (>= 1.5.0),
    purrr (>= 0.2.5),
    rlang (>= 0.2.1),
    rmarkdown (>= 1.2),
    RPostgreSQL (>= 0.4.1),
    stats (>= 3.3.0),
    stringr (>= 1.2.0),
    tidyr (>= 0.6.1),
    utils (>= 3.3.0)
    fs (>= 1.2.2),
    knitr (>= 1.15.1),
    maps (>= 3.3.0),
    mapproj (>= 1.2.6),
    testthat(>= 2.0.0),
    covr (>= 3.1.0)
```

https://github.com/AldoCompagnoni/popler

`popler` is a data retrieval package. It allows to identify and download population or community time-series data with specified characteristics (e.g. temporal and spatial replication, taxonomic coverage, ect.).

Population and community ecologists.

Yes, the rdataretriever, which similarly to `popler` facilitates discovering and downloading ecological data. `popler` differs from the rdataretriever in its focus on population and community level time-series. First, `popler` provides a database structure designed to accommodate all population and community time-series data (see database description in the manuscript, https://github.com/texmiller/popler-ms/blob/master/popler_ms.pdf). Its future growth could make it the to-go database for synthesis research on ecological time-series data. Second, `popler` focuses specifically on data from the USA Long-Term Ecological Research Netowork (LTER), particularly the data provided in the LTER data portal (https://portal.lternet.edu). The more general target of the rdataretriever does not make it a comprehensive source of LTER time-series ecological data (a handful of LTER time series datasets were provided in fall 2016).

N/A

- [x ] does not violate the Terms of Service of any service it interacts with.
- [x ] has a CRAN and OSI accepted license.
- [x ] contains a README with instructions for installing the development version.
- [x ] includes documentation with examples for all functions.
- [x ] contains a vignette with examples of its essential functions and uses.
- [x ] has a test suite.
- [x ] has continuous integration, including reporting of test coverage, using services such as Travis CI, Coveralls and/or CodeCov.
- [x ] I agree to abide by [ROpenSci's Code of Conduct](https://github.com/ropensci/onboarding/blob/master/policies.md#code-of-conduct) during the review process and in maintaining my package should it be accepted.

- [x ] Do you intend for this package to go on CRAN?
- [x ] Do you wish to submit an Applications Article about your package to [Methods in Ecology and Evolution](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/)? If so:
    - [x ] The package is novel and will be of interest to the broad readership of the journal.
    - [x ] The manuscript describing the package is no longer than 3000 words.
    - [x ] You intend to archive the code for the package in a long-term repository which meets the requirements of the journal (see [MEE's Policy on Publishing Code](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/journal-resources/policy-on-publishing-code.html))

- [x ] Does `R CMD check` (or `devtools::check()`) succeed?  Paste and describe any errors or warnings:

- [x ] Does the package conform to [rOpenSci packaging guidelines](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md)? Please describe any exceptions:

N/A

N/A
