### 2018-12-04 7a80a1f
Download summary table through API
