+++ R/util.R
@@ -123,6 +123,79 @@ factor_to_character <- function(x, full_tbl = FALSE){
   
 }
 
+# service functions that 'talk' to API -----------------------------------------
+
+popler_base <- "https://popler.space"
+cc          <- function(l) Filter(Negate(is.null), l)
+
+# get actual data from the API
+pop_GET <- function(path, args, ...) {
+  cli <- crul::HttpClient$new(url = popler_base, opts = list(...))
+  res <- cli$get(path = path, query = cc(args))
+  res$raise_for_status()
+  txt <- res$parse("UTF-8")
+  dat <- jsonlite::fromJSON(txt)
+  dat$data <- tibble::as_tibble(dat$data)
+  return(dat)
+}
+
+#' search 
+#' 
+#' @export
+#' @param limit number of records to return, default: 10
+#' @param offset record number to start at, default: first record
+#' @param ... curl options passed on to [crul::HttpClient]
+#' @examples
+#' # basic example
+#' pop_summary()
+#' # pass in curl options for debugging, seeing http request details
+#' pop_summary(verbose = TRUE)
+pop_summary <- function(limit = 10, offset = 0, ...) {
+  args <- list(limit = limit, offset = offset)
+  pop_GET("summary", args, ...)
+}
+
+#' search 
+#' 
+#' @export
+#' @param proj_metadata_key project metadat key
+#' @param limit number of records to return, default: 10
+#' @param offset record number to start at, default: first record
+#' @param ... curl options passed on to [crul::HttpClient]
+#' @examples
+#' # basic example
+#' pop_search(proj_metadata_key = 13)
+#' # pass in curl options for debugging, seeing http request details
+#' pop_search(proj_metadata_key = 13, verbose = TRUE)
+pop_search <- function(proj_metadata_key, limit = 10, offset = 0, ...) {
+  args <- list(proj_metadata_key = proj_metadata_key, 
+    limit = limit, offset = offset)
+  pop_GET("search", args, ...)
+}
+
+#' @noRd
+# set offsets and limits to download whole datasets through the API
+# (The API only downloads 1000 rows at a time)
+offset_limit <- function( route_function ){
+  
+  # count the number of rows in the dataset
+  count_summ  <- route_function( limit = 10, offset = 0 )$count
+  
+  # total number of "offsets" and "limits"
+  n_download  <- count_summ %/% 1000
+  rest        <- count_summ %% 1000
+  
+  # "offsets" and "limits"
+  download_v  <- 1000 * c(0:n_download )
+  limit_v     <- rep(1000, length(download_v) )
+
+  if( rest != 0 ) limit_v[length(download_v)] <- rest
+  
+  list(offset_v   = download_v,
+       limit_v    = limit_v )
+
+}
+
 # generate main data table summary ---------------------------------------------
 
 #' @title Update \code{popler}'s summary table
@@ -143,27 +216,88 @@ pplr_summary_table_update <- function(){
   
   message("Please wait while popler updates its summary table... this may take several minutes.")
   
-  # set database connection
-  conn <- db_open()
+  # set up list of variables to retain
   
   # list all columns
-  proj_cols   <- query_get(conn, "SELECT column_name FROM information_schema.columns WHERE table_name = 'project_table'")[,1]
-  lter_cols   <- query_get(conn, "SELECT column_name FROM information_schema.columns WHERE table_name = 'lter_table'")[,1]
-  taxa_cols   <- query_get(conn, "SELECT column_name FROM information_schema.columns WHERE table_name = 'taxa_table'")[,1]
+  taxa_cols <- c("taxa_table_key","site_in_project_taxa_key", "sppcode",                 
+                "kingdom", "subkingdom", "infrakingdom",            
+                "superdivision", "division", "subdivision",             
+                "superphylum", "phylum", "subphylum",               
[…]
