+++ NAMESPACE
@@ -18,6 +18,9 @@ export(pplr_report_metadata)
 export(pplr_site_rep)
 export(pplr_site_rep_plot)
 export(pplr_summary_table_update)
+importFrom(crul,HttpClient)
+importFrom(jsonlite,fromJSON)
+importFrom(tibble,as_tibble)
 importFrom(RPostgreSQL,PostgreSQL)
 importFrom(RPostgreSQL,dbConnect)
 importFrom(RPostgreSQL,dbDisconnect)
