### 2018-12-04 6997266
Specify imports of function pop_GET
