+++ R/util.R
@@ -128,6 +128,8 @@ factor_to_character <- function(x, full_tbl = FALSE){
 popler_base <- "https://popler.space"
 cc          <- function(l) Filter(Negate(is.null), l)
 
+#' @importFrom crul jsonlite tibble
+#' @noRd
 # get actual data from the API
 pop_GET <- function(path, args, ...) {
   cli <- crul::HttpClient$new(url = popler_base, opts = list(...))
