### 2018-12-07 0eb872e
Generalize offset_limi to work with pop_summary and pop_search
