### 2018-12-07 4c03c31
Create two separate offset_limit for 'summary' and 'search'
