### 2018-12-07 4a93027
Create progress bar when downloading summary table; comment function pplr_summary_table_check()
