### 2018-12-08 abaf9f5
Make popler_query download from API
