### 2018-12-08 672f7b9
Remove vars_query()
