+++ R/get_data.R
@@ -175,80 +175,6 @@ pplr_get_data <- function(..., add_vars = NULL, subtract_vars = NULL,
 }
 
 
-#' @noRd
-# obtain all potential columns 
-vars_query <- function(conn){
-  
-  proj_sql <- paste0("SELECT column_name ",
-                     "FROM information_schema.columns ",
-                     "WHERE table_name = 'project_table'")
-  lter_sql <-  paste0("SELECT column_name ",
-                      "FROM information_schema.columns ",
-                      "WHERE table_name = 'lter_table'")
-  site_sql <- paste0("SELECT column_name ",
-                     "FROM information_schema.columns ",
-                     "WHERE table_name = 'study_site_table'")
-  s_i_p_sql <- paste0("SELECT column_name ",
-                      "FROM information_schema.columns ",
-                      "WHERE table_name = 'site_in_project_table'")
-  taxa_sql <- paste0("SELECT column_name ",
-                     "FROM information_schema.columns ",
-                     "WHERE table_name = 'taxa_table'")
-  abund_sql <- paste0("SELECT column_name ",
-                      "FROM information_schema.columns ",
-                      "WHERE table_name = 'count_table'")
-  #list variables from the 6 tables relevant to standard popler queries  
-  proj_vars <- query_get(conn,
-                         proj_sql)[ ,1]
-  lter_vars <- query_get(conn, 
-                         lter_sql)[ ,1]
-  site_vars <- query_get(conn, 
-                         site_sql)[ ,1]
-  s_i_p_vars <- query_get(conn, 
-                          s_i_p_sql)[ ,1]
-  taxa_vars <- query_get(conn, 
-                         taxa_sql)[ ,1]
-  abund_vars <- query_get(conn, 
-                          abund_sql)[ ,1]
-  
-  # a vector containing all variables
-  all_vars <- c(proj_vars, lter_vars,
-                site_vars, s_i_p_vars,
-                taxa_vars, abund_vars)
-  
-  # remove some variables that are in the database but we don't want to return
-  # this is a temporary fix until we remove those columns from the database.
-  all_vars <- all_vars[!all_vars %in% c("currently_funded",
-                                        "homepage",
-                                        "current_principle_investigator")]
-  
-  # a vector of "default" variables
-  default_vars  <- c("authors", "authors_contact",
-                     "year", "day", "month",
-                     "sppcode", "genus", "species", "datatype",
-                     "spatial_replication_level_1_label",
-                     "spatial_replication_level_1",
-                     "spatial_replication_level_2_label",
-                     "spatial_replication_level_2",
-                     "spatial_replication_level_3_label",
-                     "spatial_replication_level_3",
-                     "spatial_replication_level_4_label",
-                     "spatial_replication_level_4",
-                     "spatial_replication_level_5_label",
-                     "spatial_replication_level_5",
-                     "proj_metadata_key",
-                     "structure_type_1", "structure_type_2",
-                     "structure_type_3", "structure_type_4",
-                     "treatment_type_1", "treatment_type_2",
-                     "treatment_type_3",
-                     "covariates" 
-  )
-  
-  return(list(all_vars = all_vars, default_vars = default_vars))
-  
-}
-
-
 #' @importFrom lazyeval lazy_dots
 #' @noRd
 # a function to concatenate browse() outputs and new arguments
