### 2018-12-08 68c3a44
Remove db_open() and db_close()
