### 2018-12-08 da10c33
Remove add/subtract_variable functionality
