### 2018-12-08 8591484
Allow pplr_query() for multiple datasets
