### 2018-12-09 f778f65
Allow three ways to use get_data
