### 2018-12-10 501e3c1
Fix issues downloading entire data sets
