+++ NAMESPACE
@@ -24,11 +24,13 @@ importFrom(tibble,as_tibble)
 importFrom(RPostgreSQL,PostgreSQL)
 importFrom(RPostgreSQL,dbConnect)
 importFrom(RPostgreSQL,dbDisconnect)
+importFrom(plyr,colwise)
 importFrom(dbplyr,sql)
 importFrom(dbplyr,translate_sql)
 importFrom(dbplyr,translate_sql_)
 importFrom(dplyr,"%>%")
 importFrom(dplyr,bind_cols)
+importFrom(dplyr,bind_rows)
 importFrom(dplyr,filter)
 importFrom(dplyr,group_by)
 importFrom(dplyr,mutate)
