### 2018-12-10 423b79c
Point out important characteristics of pplr_get_data()
