### 2018-12-10 8aab908
Set DATE and OBSERVATION to numeric
