### 2018-12-10 7d0f22f
Update example using pplr_site_rep()
