### 2018-12-10 53ff881
Delete tests on get_data() utility functions
