### 2018-12-16 9bb1c44
Fix 'summary_table' bug in pplr_get_data
