+++ DESCRIPTION
@@ -30,13 +30,14 @@ Imports:
     utils (>= 3.3.0)
 License: MIT + file LICENSE
 Encoding: UTF-8
-Suggests: 
+Suggests:
     fs (>= 1.2.2),
     knitr (>= 1.15.1),
     maps (>= 3.3.0),
     mapproj (>= 1.2.6),
     testthat(>= 2.0.0),
-    covr (>= 3.1.0)
+    covr (>= 3.1.0),
+    rmarkdown
 LazyData: true
 RoxygenNote: 6.1.0
 VignetteBuilder: knitr
