### 2018-12-17 5c4811a
Remove function and data linked to DB
