### 2018-12-17 6808b45
Remove obsolete comment!
