### 2018-12-17 4b68b62
Remove obsolete demo files
