### 2018-12-17 5888ecf
Include branch api in travis build
