### 2019-01-21 f494517
Open multiple urls per each study (specific to 4 GCE studies)
