### 2019-01-21 1af20e6
Notify how many links are associated with each study
