### 2019-01-21 5c863a3
Install libgit2 explicitly
