### 2019-01-22 c259158
Exclude a standard build of osx, oldrel
