+++ DESCRIPTION
@@ -13,20 +13,24 @@ URL: https://github.com/AldoCompagnoni/popler
 BugReports: https://github.com/AldoCompagnoni/popler/issues
 Depends: R (>= 3.0.0)
 Imports:
+    curl (>= 0.5.2),
     dbplyr (>= 1.1.0), 
     dplyr (>= 0.5.0), 
     ggplot2 (>= 3.0.0),
     grid (>= 3.3.0),
+    jsonlite (>= 1.5),
     lazyeval (>= 0.2.0),
     lubridate (>= 1.7.4),
     magrittr (>= 1.5.0),
+    plyr (>= 1.8.4),
     purrr (>= 0.2.5),
     rlang (>= 0.2.1),
     rmarkdown (>= 1.2),
     RPostgreSQL (>= 0.4.1),
     stats (>= 3.3.0),
     stringr (>= 1.2.0),
-    tidyr (>= 0.6.1),  
+    tibble (>= 1.4.2),
+    tidyr (>= 0.6.1),
     utils (>= 3.3.0)
 License: MIT + file LICENSE
 Encoding: UTF-8
@@ -36,8 +40,7 @@ Suggests:
     maps (>= 3.3.0),
     mapproj (>= 1.2.6),
     testthat(>= 2.0.0),
-    covr (>= 3.1.0),
-    rmarkdown
+    covr (>= 3.1.0)
 LazyData: true
 RoxygenNote: 6.1.0
 VignetteBuilder: knitr
