### 2019-01-22 3ecf40c
Fix reporting and cov_unpack tests
