+++ tests/testthat/test-reporting.R
@@ -4,7 +4,7 @@ context('test report_dictionary')
 library(fs)
 library(ggplot2)
 
-data_obj <- pplr_get_data( proj_metadata_key == 11 )
+data_obj <- pplr_get_data( proj_metadata_key == 1 )
 
 test_that('Output file types are correct', {
   skip_on_cran()
