### 2019-01-23 2f36aef
fix rooxygen typos and re-build docs
