+++ NAMESPACE
@@ -6,6 +6,8 @@ S3method(mutate,browse)
 S3method(mutate,get_data)
 export(filter)
 export(mutate)
+export(pop_search)
+export(pop_summary)
 export(pplr_browse)
 export(pplr_citation)
 export(pplr_cov_unpack)
@@ -18,14 +20,9 @@ export(pplr_report_metadata)
 export(pplr_site_rep)
 export(pplr_site_rep_plot)
 export(pplr_summary_table_update)
-importFrom(crul,HttpClient)
-importFrom(jsonlite,fromJSON)
-importFrom(tibble,as_tibble)
-importFrom(RPostgreSQL,PostgreSQL)
-importFrom(RPostgreSQL,dbConnect)
-importFrom(RPostgreSQL,dbDisconnect)
-importFrom(plyr,colwise)
-importFrom(dbplyr,sql)
+import(crul)
+import(jsonlite)
+import(tibble)
 importFrom(dbplyr,translate_sql)
 importFrom(dbplyr,translate_sql_)
 importFrom(dplyr,"%>%")
@@ -37,7 +34,6 @@ importFrom(dplyr,mutate)
 importFrom(dplyr,n)
 importFrom(dplyr,select)
 importFrom(dplyr,summarise)
-importFrom(dplyr,tbl)
 importFrom(dplyr,tbl_df)
 importFrom(dplyr,tibble)
 importFrom(dplyr,ungroup)
