### 2019-01-23 #254 @noamross
### Editor checks:

---

#### Editor comments

Thanks for your (restarted) submission!

Here is `goodpractice::gp()` output.  I see a few NOTES that should be fixed.

```
── GP popler ───────────────────────────────

It is good practice to

    R/browse.R:58:NA
    R/browse.R:59:NA
    R/browse.R:60:NA
    R/browse.R:61:NA
    R/browse.R:62:NA
    ... and 1062 more lines

  ✖ not import packages as a whole, as this can cause name
    clashes between the imported packages. Instead, import only the
    specific functions you need.
  ✖ fix this R CMD check NOTE: Namespace in Imports field not
    imported from: ‘RPostgreSQL’ All declared Imports should be used.
  ✖ fix this R CMD check NOTE: pplr_get_data: no visible
    binding for global variable ‘.’
    (/private/var/folders/82/07j7dctj31lg1tw6w1c_fdw40000gn/T/RtmplJK6qV/file104d3edc4f8e/popler.Rcheck/00_pkg_src/popler/R/get_data.R:87-90)
    pplr_get_data: no visible binding for global variable ‘year’
    (/private/var/folders/82/07j7dctj31lg1tw6w1c_fdw40000gn/T/RtmplJK6qV/file104d3edc4f8e/popler.Rcheck/00_pkg_src/popler/R/get_data.R:104-107)
    pplr_get_data: no visible binding for global variable ‘month’
    (/private/var/folders/82/07j7dctj31lg1tw6w1c_fdw40000gn/T/RtmplJK6qV/file104d3edc4f8e/popler.Rcheck/00_pkg_src/popler/R/get_data.R:104-107)
    pplr_get_data: no visible binding for global variable ‘day’
    (/private/var/folders/82/07j7dctj31lg1tw6w1c_fdw40000gn/T/RtmplJK6qV/file104d3edc4f8e/popler.Rcheck/00_pkg_src/popler/R/get_data.R:104-107)
    pplr_query: no visible binding for global variable ‘.’
    (/private/var/folders/82/07j7dctj31lg1tw6w1c_fdw40000gn/T/RtmplJK6qV/file104d3edc4f8e/popler.Rcheck/00_pkg_src/popler/R/get_data.R:257-261)
    pplr_query: no visible global function definition for
    ‘txtProgressBar’
    (/private/var/folders/82/07j7dctj31lg1tw6w1c_fdw40000gn/T/RtmplJK6qV/file104d3edc4f8e/popler.Rcheck/00_pkg_src/popler/R/get_data.R:269)
    pplr_query : downld_dataset: no visible global function definition
    for ‘setTxtProgressBar’
    (/private/var/folders/82/07j7dctj31lg1tw6w1c_fdw40000gn/T/RtmplJK6qV/file104d3edc4f8e/popler.Rcheck/00_pkg_src/popler/R/get_data.R:273)
    pplr_query: no visible binding for global variable ‘.’
    (/private/var/folders/82/07j7dctj31lg1tw6w1c_fdw40000gn/T/RtmplJK6qV/file104d3edc4f8e/popler.Rcheck/00_pkg_src/popler/R/get_data.R:285-287)
    pplr_summary_table_update: no visible global function definition
    for ‘txtProgressBar’
    (/private/var/folders/82/07j7dctj31lg1tw6w1c_fdw40000gn/T/RtmplJK6qV/file104d3edc4f8e/popler.Rcheck/00_pkg_src/popler/R/util.R:322)
    pplr_summary_table_update : downld_summary: no visible global
    function definition for ‘setTxtProgressBar’
    (/private/var/folders/82/07j7dctj31lg1tw6w1c_fdw40000gn/T/RtmplJK6qV/file104d3edc4f8e/popler.Rcheck/00_pkg_src/popler/R/util.R:326)
    Undefined global functions or variables: . day month
    setTxtProgressBar txtProgressBar year Consider adding
    importFrom("utils", "setTxtProgressBar", "txtProgressBar") to your
    NAMESPACE file.
  ✖ avoid 'T' and 'F', as they are just variables which are
    set to the logicals 'TRUE' and 'FALSE' by default, but are not
    reserved words and hence can be overwritten by the user.  Hence,
    one should always use 'TRUE' and 'FALSE' for the logicals.

    R/util.R:NA:NA
```

These notes are not major and can be fixed while I seek reviewers.  A see a few are due to using non-standard evaluation inside functions.  These should be able to be dealt with by quoting arguments and avoiding the `.` construction, instead using `dplyr::pull`

Test coverage is actually 76%, as reported by CodeCov (the reported number above is from run `--as-cran`). This is OK.  I will instruct reviewers to examine coverage reports and check that uncovered code sufficiently straightforward to not need testing.

Please also run `spelling::spell_check_package()` and look for true typos.

---

Reviewers:
Due date:
