### 2019-02-13 #254 @cgries
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

not sure there are any claims and downloads are very slow

---

This R package provides access to a great data resource for population ecologists. The authors compiled a large number of datasets into a standardized data format with standardized metadata on which the user can query. This effort will open the door to important meta-analyses of data from the NSF Long-term Ecological Research Network.

README:
Although there is one sentence about what this package is doing, there are no further information about the actual database or the research applications this is built for. If this is the first point of entry for someone not familiar with the Popler project it would be very hard to judge what this package does.
There are no links to the vignettes. However, important information on the actual data is only available in the vignettes.
The vignettes are well written and help understand the intent and functionality of the functions and structure of the databsae, the latter I was missing in the online information available before download.
Somewhere I came across a publication on Popler, but I can't find the link anymore. Would be great to have that in the README.

To really test out the capabilities, I would need to know much more about the database itself. What do all those metadata mean that I could use to subset the data? E.g., what are structure_types 'types of indidivual structure' 1 - 880?

the pplr_report_metadata() and pplr_browse(report=TRUE) functions are pretty slick and produce a very nice display of metadata

Testing examples:

SEV_studies <- pplr_get_data(lterid == 'SEV')
Error: Bad Request (HTTP 400)

BNZ <- pplr_get_data(lterid == 'BNZ')
Error: Bad Request (HTTP 400)

when it works data download is really slow:

poa_data <- pplr_get_data(poa_metadata)
took almost 10 minutes to complete on a pretty fast university connection

penguin_raw_data <- pplr_get_data(penguins, cov_unpack = TRUE)
took over 15 minutes

Antarctica <- pplr_get_data(lterid == 'PAL')
took about 15 minutes

just a curious observation:
test_1 <- pplr_browse(proj_metadata_key == 133)
test_1_raw_data <- pplr_get_data(test_1)
produces a dataset where the author column is completley filled with 'Sherry Johnson'

while pplr_citation(test_1_raw_data) produces this citation
$bibliography
Farrand M (2014). “Aquatic insect sampling in Lookout Creek at the H.J. Andrews
Experimental Forest, 2001. Environmental Data Initiative.” Andrew Forest LTER, <URL:
http://dx.doi.org/10.6073/pasta/babf56fed5882756432f28982312a551>.

## devtools::test()

Testing popler
ok | OK F W S | Context
ok |  4       | browse(): Informational messages
ok |  3       | browse(): Select by criteria
ok | 13       | browse(): taxa_nest [2.4 s]
ok |  7       | pplr_browse(): error functions
ok | 13       | browse() function [8.4 s]
x  |  0 1     | cov_unpack [141.9 s]
ok | 12       | dictionary() function [4.5 s]
x  |  0 1     | dplyr verbs [472.2 s]
ok | 12       | lter_map functions [29.3 s]
ok |  5       | popler_citation [6.8 s]
x  |  0 1     | test report_dictionary [131.7 s]
x  |  0 1     | pplr_site_rep [13.3 s]

== Results
OK:       69
1: object ‘jsonlite’ is not exported by 'namespace:crul'
2: object ‘tibble’ is not exported by 'namespace:crul'

## devtools::check()

failed for me, but that may be my system

## goodpractice::gp()

 GP popler

   avoid 'T' and 'F', as they are just variables which are set to the logicals
    'TRUE' and 'FALSE' by default, but are not reserved words and hence can be overwritten
    by the user.  Hence, one should always use 'TRUE' and 'FALSE' for the logicals.

    R/util.R:NA:NA
