### 2019-02-15 #254 @bpbond
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

3
---

Thanks for the effort you put into this package @AldoCompagnoni – it has a lot of potential and many parts are very well done.

Preliminaries: I agree with (and so won't repeat in detail) @cgries 's comments about the lack of easy documentation about the popler database itself, research applications, and other background information; the extremely slow downloads; and generally good vignettes. Here are some other
issues that jumped out at me while evaluating:

#### Dependencies

`popler` has a lot of dependencies. In particular, I see it imports both `plyr` and `dplyr`
...is this necessary, in particular the pretty old and (I think) no-longer-updated `plyr`
that is only used (as far as I see) in one place?

#### Duplication of functionality

I personally like the philosophy that packages should be as simple and focused as possible.
From that point of view, I don't see that
```R
pplr_browse(duration_years > 20)
```

gives any advantage over
```R
pplr_browse() %>% filter(duration_years > 20)
```

In other words, why duplicate the `dplyr::filter` functionality in `pplr_browse()`? (A similar argument applies to its `vars` parameter.) This is particularly true since you're already providing dplyr method overrides for some of the package output classes.

...is OK, but many minor grammatical mistakes and incomplete things. For example, the
documentation for `pplr_browse()` says: "keyword	  A string that selects". This is not very informative. In this case, at least note that `keyword` is a regular expression that is passed to `grepl` and used to search across all fields.

#### Code style

The `popler` code is generally clear, consistent, and easy to follow. In-code comments are minimal but usually useful when present.

The tests take a long time to run, and when `NOT_CRAN` is TRUE, a _super_ long time. (So long that I never could look closely at the coverage reports as @noamross requested above.) I assume this is because of the slow downloading issues. It would be great if most tests used network [stubs](https://cran.r-project.org/web/packages/stubthat/vignettes/stubthat.html) so that tests run quickly–this will make them much more useful!

#### Report capability

The report generation of `pplr_browse()` is neat and potentially very useful.
But the console output from the example is kind of scary:
```R
> pplr_browse(report=TRUE)
NOTE! Studies 297 are linked to a total of58 URLsNOTE! Studies 298 are linked to a total of58 URLsNOTE! Studies 299 are linked to a total of58 URLsNOTE! Studies 300 are linked to a total of58 URLsNOTE! Studies 313 are linked to a total of58 URLs
NOTE! Studies 297 are linked to a total of58 URLsNOTE! Studies 298 are linked to a total of58 URLsNOTE! Studies 299 are linked to a total of58 URLsNOTE! Studies 300 are linked to a total of58 URLsNOTE! Studies 313 are linked to a total of58 URLs
...(etc)...
In withCallingHandlers(.External2(C_parseRd, tcon, srcfile, "UTF-8",  :
  <connection>:4: unexpected END_OF_INPUT '.
'
```
Can this be improved?

**In summary**, I greatly appreciate the goals of `popler`. I do think it needs better documentation about what the underlying database is; to be simplified in terms of dependencies and functionality; network stubs for the tests; and improved documentation in many places.
