### 2019-02-22 #254 @AldoCompagnoni
Dear all,
Thank you for the very useful feedback!

I have identified several points on which to focus our work in the next few weeks:

1. There needs to be the ABC on the database structure, probably in a separate Vignette, or in an existing Vignette.
We're already working on this, and we estimate uploading the schema and description by March 10th.
To address a comment by Corinna, if you are curious to know how the database works, a description of the package is already available online as part of the to-be-submitted article connected to popler:
https://github.com/texmiller/popler-ms/blob/master/popler_ms.pdf

2. Downloads are too long.
This is a problem caused by the combination of API working on an Amazon instance with very little Ram (1GB). There are two ways to address this issue:
First, upgrading to an instance with more RAM will speed things up. Unfortunately, however, I do not have an estimate of how much faster the queries would become.
Second, we could change the main query we use to fetch data on popler. During development of the API, I developed a query using the SQL clause "WITH" which sped thing up by a factor 10. However, Scott Chamberlain - and anybody else I know - do not know how to implement this SQL query in Ruby, which is the language in which the API is programmed.
Of course I lean towards solution two, and I am actively asking everybody I know to solve this issue. Little success so far, but a couple of weeks of time will tell!

3. Documentation files have grammatical mistakes/imprecise language:
This will be a simple fix.

4. We need to work on how to handle URLs/citations:
I realize that we need a better way to handle the multiple-URL datasets provided by the Georgia Coastal Ecosystems LTER. This came out prominently in the comment by Bob on "Report capability". So far, I might remove the warnings when using `popler_browse(report=T)`, and I might prompt the user for a "yes" or "no" when using `pplr_metadata_url()`
Also, there seem to be a few apparent mistakes in citations - thank you for pointing those out. This is another simple fix.

5. Finally, Bob suggested to substitute

`pplr_browse(duration_years > 20)`
with
`pplr_browse() %>% filter(duration_years > 20)`

The reason that we use `pplr_browse(duration_years > 20)`, is that if we store `pplr_browse(duration_years > 20)` in an object, this object can directly be passed on to `pplr_get_data()` for download. This is not possible with an object produced via the synthax `pplr_browse() %>% filter(duration_years > 20)`. Admittedly, at the top of my head I cannot answer whether it is possible to retain the same functionality using the `pplr_browse() %>% filter(duration_years > 20)` synthax, but I am currently working on finding a reliable answer!
