+++ NAMESPACE
@@ -21,9 +21,6 @@ export(pplr_summary_table_update)
 importFrom(crul,HttpClient)
 importFrom(jsonlite,fromJSON)
 importFrom(tibble,as_tibble)
-importFrom(RPostgreSQL,PostgreSQL)
-importFrom(RPostgreSQL,dbConnect)
-importFrom(RPostgreSQL,dbDisconnect)
 importFrom(plyr,colwise)
 importFrom(dbplyr,sql)
 importFrom(dbplyr,translate_sql)
