+++ vignettes/popler-database.Rmd
@@ -1,0 +1,149 @@
+---
+title: "Popler Database"
+author: "Andrew Bibian"
+date: "3/16/2019"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Put the title of your vignette here}
+  %\VignetteEngine{knitr::rmarkdown}
+  \usepackage[utf8]{inputenc}
+---
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE)
+```
+
+## Popler Overview
+Popler is a centralized database for population dynamics datasets gathered throughout 
+NSF's Long Term Ecological Resarch (LTER) stations. 
+
+We define population dynamics datasets as follows:
+
+- Time series data of a randomly sampled population measuring changes in abundance, 
+density, biomass, cover class or individual measures on a temporally regular 
+interval over a standardized set of spatial units and sampling effort. 
+The interval of time must encompass the life cycle of the organism such that 
+birth and deaths occur during the interval.
+
+### Temporal Referencing
+All temporal information within a dataset is stored using exact 
+sampling dates up to the day resolution, when available. Since not all datasets 
+have the same temporal resolution Popler stores date time information in three 
+separate columns; **day, month, and year**. In any of the date time columns, 
+`NULL`  or `-99999` values indicate that the information was not available 
+from the raw data. 
+
+**Never do we perform any temporal aggregations of data prior to storing them in Popler.**
+
+### Spatial Referencing
+We have chosen to reference sampling locations (spatial levels) by the unit of replication 
+a sample represents in an experimental design. 
+
+For example, a study performed at 5 sites, with 3 transects at each site, and 4 
+quadrats within each transect is referenced as having 3 levels of spatial 
+replication:
+
+- Level 1: the site level, 
+- Level 2: the transect level
+- Level 3: the quadrat level
+
+This referencing scheme allows us to standardize and align datasets collected 
+from a variety of different sampling designs and across different data types. 
+
+Metadata related to the extent of a spatial sampling unit is recorded when 
+available (i.e. km2, m2, cm2, etc.). Note, higher levels of spatial replication 
+indicate smaller areas of sampling extent. 
+
+
+**Never do we aggregate data by levels of spatial replication prior to storing them in Popler.**
+
+#### Spatial Replication Level 1
+By referencing sampling locations in the above manner a ‘site’ label within a dataset
+represent the highest level of spatial replication (level 1).  With permanent 'sites' 
+existing at LTER stations and reused across studies, ‘site’ lables 
+allow querying of data collected for particular spatial locations regardless of the 
+original study a dataset was derived from.
+
+When possible, latitude and longitude coordinates are also recorded if available in 
+the source metadata; -99999 values exist within these fields for studies that do not 
+record geographic location for their 'site' label. 
+
+
+## The Popler Database Schema
+
+The following depicts the relationship structure among 14 tables. Below we will
+discuss this schema, table contents, and give definitions. Three
+tables will not be discussed as two are climate related data tables yet to be populated
+and one is a table for database migration information.
+
+```{r, out.width = "675px", echo=FALSE}
+knitr::include_graphics("img/full-schema.png")
+```
+
+## 1. LTER Table
+
+```{r, out.width = "600px", echo=FALSE}
+knitr::include_graphics("img/lter-table.png")
+```
+
+## 2. Study Site Table
+
[…]
