### 2019-03-24 1f94d74
Update markdown for table descirptions and metadata info
