+++ data-raw/internal_data.R
@@ -1,4 +1,6 @@
 
+# explanations ---------------------------------------------------------
+
 #' @noRd
 # data frame of variable explanations
 explanations = data.frame(variable = c('proj_metadata_key',
@@ -147,7 +149,154 @@ explain_short = data.frame( variable = c('title','proj_metadata_key',
                                           "family","genus"),
                             stringsAsFactors = F)
 
+
+# data for tests -------------------------------------------------------
+
+ 
+# format data 
+download_popler_test_data <- function( proj_id ){
+  
+  # all service functions 
+  
+  # connect to the database
+  popler_connector <- function(dbname=NULL, host=NULL, port=NULL, user=NULL, password=NULL, silent=TRUE) {
+  
+    if (!requireNamespace("RPostgreSQL", quietly = TRUE)) {
+      stop("RPostgreSQL package required to connect to postgres db", call. = FALSE)
+    }
+    user <- if(is.null(user)){
+      if(identical(Sys.getenv("TRAVIS"), "true")){"postgres"} else {""}
+      } else user
+    con <- RPostgreSQL::dbConnect(RPostgreSQL::PostgreSQL(),
+                     host     = if(is.null(host))     "" else host,
+                     dbname   = if(is.null(dbname))   "" else dbname,
+                     user     = user,
+                     password = if(is.null(password)) "" else password,
+                     port     = if(is.null(port))     "" else port)
+  }
+  
+  # database open function
+  db_open <- function(){
+    popler_connector(dbname="",
+                      host="",
+                      port=,
+                      user="",
+                      password="",
+                      silent=TRUE)
+  }
+  
+  # get query 
+  query_get <- function(connection, query){
+    # accepts a connection and a string query input
+    # outputs a dataframe
+    return(dplyr::tbl(connection, dbplyr::sql(query)) %>% data.frame())
+  }
+  
+  # open connection
+  conn      <- db_open()
+  
+  # produce efficient query (uses WITH clause)
+  efficienty_query <- function( proj_id ){
+    
+    paste0('
+    WITH project_taxa_table AS (
+         select authors, authors_contact,   -- 126
+                datatype,
+                spatial_replication_level_1_label, spatial_replication_level_2_label, spatial_replication_level_3_label, spatial_replication_level_4_label, spatial_replication_level_5_label, proj_metadata_key, structured_type_1, structured_type_2, structured_type_3, structured_type_4,
+                sppcode, genus, species, taxa_table.taxa_table_key
+         FROM   taxa_table            
+           JOIN site_in_project_table ON taxa_table.site_in_project_taxa_key =site_in_project_table.site_in_project_key 
+           JOIN project_table         ON site_in_project_table.project_table_fkey = project_table.proj_metadata_key 
+           JOIN study_site_table      ON site_in_project_table.study_site_table_fkey =study_site_table.study_site_key 
+           JOIN lter_table            ON study_site_table.lter_table_fkey = lter_table.lterid 
+    WHERE proj_metadata_key = ',proj_id,' )
+    
+    SELECT authors, authors_contact, year, day, month, sppcode, genus, species, datatype, spatial_replication_level_1_label, spatial_replication_level_1, spatial_replication_level_2_label, spatial_replication_level_2, spatial_replication_level_3_label, spatial_replication_level_3, spatial_replication_level_4_label, spatial_replication_level_4, spatial_replication_level_5_label, spatial_replication_level_5, proj_metadata_key, structure_type_1, structure_type_2, structure_type_3, structure_type_4, count_table.treatment_type_1, count_table.treatment_type_2, count_table.treatment_type_3, covariates , count_observation 
+    FROM project_taxa_table
[…]
