+++ vignettes/popler-database.Rmd
@@ -35,34 +35,26 @@ knitr::opts_chunk$set(message=FALSE, warning=FALSE)
 ```
 
 ## Popler Overview
-Popler is a centralized database for population dynamics datasets gathered throughout 
+Popler is a database of population and individual-level data gathered throughout 
 NSF's Long Term Ecological Resarch (LTER) stations. 
 
-We define population dynamics datasets as follows:
+We define population data datasets as time series of the size or density of a population of a taxonomic unit. The size of a population can be quantified as a count, biomass, or cover class. These measures are always repeated temporally and often repeated spatially.
+We define individual-level data as information on the attributes of the individuals, or a subset thereof, that make up a population.
 
-- Time series data of a randomly sampled population measuring changes in abundance, 
-density, biomass, cover class or individual measures on a temporally regular 
-interval over a standardized set of spatial units and sampling effort. 
-The interval of time must encompass the life cycle of the organism such that 
-birth and deaths occur during the interval.
 
-### Temporal Referencing
-All temporal information within a dataset is stored using exact 
-sampling dates up to the day resolution, when available. Since not all datasets 
-have the same temporal resolution Popler stores date time information in three 
+### Temporal replication
+All temporal information within a dataset is stored using 
+sampling dates up to the daily resolution, when available. Since not all datasets have the same temporal resolution, popler stores date time information in three 
 separate columns; **day, month, and year**. In any of the date time columns, 
 `NULL`  or `-99999` values indicate that the information was not available 
 from the raw data. 
 
 **Never do we perform any temporal aggregations of data prior to storing them in Popler.**
 
-### Spatial Referencing
-We have chosen to reference sampling locations (spatial levels) by the unit of replication 
-a sample represents in an experimental design. 
+### Spatial replication
+We have chosen to reference sampling locations (spatial levels) by the unit of replication a sample represents in an experimental design. 
 
-For example, a study performed at 5 sites, with 3 transects at each site, and 4 
-quadrats within each transect is referenced as having 3 levels of spatial 
-replication:
+Most datasets contain spatially nested data. For example, a study performed at 5 sites, with 3 transects at each site, and 4 quadrats within each transect is referenced as having 3 levels of nested spatial replication:
 
 - Level 1: the site level, 
 - Level 2: the transect level
@@ -71,37 +63,29 @@ replication:
 This referencing scheme allows us to standardize and align datasets collected 
 from a variety of different sampling designs and across different data types. 
 
-Metadata related to the extent of a spatial sampling unit is recorded when 
-available (i.e. km2, m2, cm2, etc.). Note, higher levels of spatial replication 
+Metadata on the extent of each spatial sampling unit is recorded when 
+available (i.e. km2, m, cm3, etc.). Note, higher levels of spatial replication 
 indicate smaller areas of sampling extent. 
 
 **Never do we aggregate data by levels of spatial replication prior to storing them in Popler.**
 
 #### Spatial Replication Level 1
-By referencing sampling locations in the above manner a ‘site’ label within a dataset
-represent the highest level of spatial replication (level 1).  With permanent 'sites' 
-existing at LTER stations and reused across studies, ‘site’ lables 
-allow querying of data collected for particular spatial locations regardless of the 
-original study a dataset was derived from.
+The first level of spatial replication is also the highest level of nested spatial replication. We call level 1 a ‘site’ within a dataset
[…]
