### 2019-04-13 518d4e5
Update documentation of functions
