+++ R/popler.R
@@ -1,7 +1,7 @@
 #' popler package
 #' 
-#' \code{popler} is a package for interacting with the PostgreSQL data base 
-#' of the same name. It contains data on long term population dynamics from
+#' \code{popler} is a package for interacting with the homonymous PostgreSQL 
+#' data base. \code{popler} contains data on long-term population dynamics from
 #' the LTER network. Every exported function is prefixed with \code{pplr_} and
 #' then a verb (e.g. \code{pplr_get_data()} or noun (e.g. \code{pplr_citation}.
 #' Accessing \code{popler} does not require an API key, all you need is an 
