+++ R/browse.R
@@ -1,17 +1,21 @@
 #' Browse the metadata of projects contained in the popler database
 #'
 #' pplr_browse() reports the metadata of LTER studies contained in the popler database. 
-#' The user can subset what data, and which variables to visualize.  
+#' The user can subset which datasets, and which metadata variables to visualize.  
 #' 
-#' @param ... A logical expression to subset popler's main table
-#' @param full_tbl Should the function return the standard 
-#' columns, or the full main table?
-#' @param vars A vector of characters: which variables 
-#' of popler's main table should be selected?
+#' @param ... A logical expression to subset the table containing the metadata of 
+#' datasets contained in popler
+#' @param full_tbl logical; Should the function returns the standard 
+#' columns, or the full main table? Default is \code{FALSE}.
+#' @param vars A vector of characters in case the user want to select 
+#' which variables of popler's main table should be selected?
 #' @param view If TRUE, opens up a spreadsheet-style data viewer.
-#' @param keyword A string that selects 
-#' @param report If TRUE, function produces a markdown 
-#' report about each study's metadata, and opens it as a html page.
+#' @param keyword A string used to select individual datasets based on
+#' pattern matching. The string is matched to every string element in the 
+#' variables of the metadata table in popler. 
+#' @param report logical; If \code{TRUE}, function produces a markdown 
+#' report about each study's metadata, and opens it as a html page. 
+#' Default is \code{FALSE}.
 #' @return A data frame combining the metadata of each project 
 #' and the taxonomic units associated with each project.
 #' @return This data frame is of class \code{popler}, \code{data.frame},
