+++ R/popler_citation.R
@@ -1,9 +1,10 @@
-#' Provide citations for a popler browse or get_data object.
+#' Provide citations for a popler object returned by \code{pplr_browse} or
+#' \code{pplr_get_data} object.
 #'
 #' Returns a bibliography, Bibtex citations, and acknowledgement template.
-#' @param input An object  of class \code{browse} or \code{get_data}.
+#' @param input An object of class \code{browse} or \code{get_data}.
 #' @param bibtex_path Specify the filename and location for 
-#' the generated markdown file (optional)
+#' the generated markdown file (optional).
 #' 
 #' @return A list of references from \code{input}.
 #' 
