### 2019-04-14 43a1289
Update grammar and text in vignettes/Readme
