+++ vignettes/popler-database-structure.Rmd
@@ -35,50 +35,43 @@ knitr::opts_chunk$set(message=FALSE, warning=FALSE)
 ```
 
 ## Popler Overview
-Popler is a database of population and individual-level data gathered throughout 
-NSF's Long Term Ecological Resarch (LTER) stations. 
+Popler is a database of population and individual-level data gathered throughout the Long Term Ecological Resarch (LTER) stations funded by the USA National Science Foundation.
 
-We define population data datasets as time series of the size or density of a population of a taxonomic unit. The size of a population can be quantified as a count, biomass, or cover class. These measures are always repeated temporally and often repeated spatially.
-We define individual-level data as information on the attributes of the individuals, or a subset thereof, that make up a population.
+We define _population data_ datasets as time series on the size or density of a population of a taxonomic unit. The size of a population can be quantified as a count, biomass, or cover class. These measures are always repeated temporally, and they are often repeated spatially as well.
+
+We define _individual-level_ data as information on the attributes of the individuals, or a subset thereof, that make up a population. For example, common attributes of individuals are size, age, and sex.
 
 
 ### Temporal replication
 All temporal information within a dataset is stored using 
-sampling dates up to the daily resolution, when available. Since not all datasets have the same temporal resolution, popler stores date time information in three 
-separate columns; **day, month, and year**. In any of the date time columns, 
-`NULL`  or `-99999` values indicate that the information was not available 
-from the raw data. 
+sampling dates up to, when available, the daily resolution. Since not all datasets have the same temporal resolution, popler stores date information in three separate columns; **day, month, and year**. In any of the date time columns, `NULL` or `-99999` values indicate that the information was not available from the raw data. 
 
-**Never do we perform any temporal aggregations of data prior to storing them in Popler.**
+Note that **never do we perform any temporal aggregations of data prior to storing them in Popler.**
 
 ### Spatial replication
-We have chosen to reference sampling locations (spatial levels) by the unit of replication a sample represents in an experimental design. 
-
-Most datasets contain spatially nested data. For example, a study performed at 5 sites, with 3 transects at each site, and 4 quadrats within each transect is referenced as having 3 levels of nested spatial replication:
+`popler` subdivides spatial replicates based on their spatial nestedness, because most sampling designs are spatially nested. For example, a study performed at 5 sites, with 3 transects at each site, and 4 quadrats within each transect, contains 3 levels of nested spatial replication:
 
 - Level 1: the site level, 
 - Level 2: the transect level
 - Level 3: the quadrat level
 
-This referencing scheme allows us to standardize and align datasets collected 
-from a variety of different sampling designs and across different data types. 
+This referencing scheme allows us to standardize and align datasets collected from a variety of different sampling designs and across different data types.
 
 Metadata on the extent of each spatial sampling unit is recorded when 
 available (i.e. km2, m, cm3, etc.). Note, higher levels of spatial replication 
 indicate smaller areas of sampling extent. 
 
-**Never do we aggregate data by levels of spatial replication prior to storing them in Popler.**
+Similarly to temporal replication, **never do we aggregate data by levels of spatial replication prior to storing them in Popler.**
 
 #### Spatial Replication Level 1
-The first level of spatial replication is also the highest level of nested spatial replication. We call level 1 a ‘site’ within a dataset
-represent the highest level of spatial replication (level 1). LTER stations contain permanent 'sites' which are reused across studies. Therefore, ‘site’ lables allow querying data collected at a particular site regardless of the original study a dataset was derived from.
[…]
