+++ vignettes/vetting-popler.Rmd
@@ -32,7 +32,7 @@ pplr_browse(duration_years > 10)
 
 ```
 
-Note that most LTER projects contemplate sampling at a yearly or sub-yearly frequency. Thus, studies longer than 10 years generally guarantee a longitudinal series of 10 or more observations. Note that the `duration_years` variable is calculated as `studyendyr - studystartyr`. Thus, an additional variable named `samplefreq` characterizes the approximate sample frequency of each study. 
+Note that most LTER projects contemplate sampling at a yearly or sub-yearly frequency. Thus, studies longer than 10 years often guarantee a longitudinal series of 10 or more observations. Note that the `duration_years` variable is calculated as `studyendyr - studystartyr`. Thus, an additional variable named `samplefreq` characterizes the approximate sample frequency of each study. 
 
 ```{r, warning = FALSE, message = FALSE}
 
@@ -41,13 +41,13 @@ pplr_browse(samplefreq == "monthly")
 
 ```
 
-Note that `samplefreq` is **not** a default variable included in the `pplr_dictionary` or `pplr_browse()` functions. This can be viewed by specifying the `full_tbl = TRUE` argument in either function.
+Note that `samplefreq` is **not** a default variable included in the `pplr_dictionary` or `pplr_browse()` functions. This can be viewed by specifying the `full_tbl = TRUE` argument in either of the above functions.
 
 ###1. Spatial replication
 
 #### Before downloading data
 
-If you wish to select data sets based on their spatial replication, you need to consider that `popler` organizes data in nested spatial levels. For example, in many plant studies data is collected at the *plot* level, which can be nested within *block*, which in turn can be nested within *site*. `popler` labels spatial levels using numbers: 1 is the coarsest level of replication which contains all other spatial replicates. In the example above, spatial level 1 is *site*, spatial level 2 is *block*, and spatial level 3 is *plot*. `popler` allows for a total of 5 spatial levels. Given the above, you can select studies based on three criteria:
+If you wish to select data sets based on their spatial replication, you need to consider that `popler` organizes data in nested spatial levels. For example, in many plant studies data is collected at the *plot* level, which can be nested within *block*, which in turn can be nested within *site*. `popler` labels spatial levels using numbers. Spatial level 1 is the coarsest level of replication which contains all other spatial replicates. In the example above, spatial level 1 is *site*, spatial level 2 is *block*, and spatial level 3 is *plot*. `popler` allows for a total of 5 spatial levels. Given the above, you can select studies based on three criteria:
 
  1. The total number of spatial replicates.
  
@@ -55,7 +55,7 @@ If you wish to select data sets based on their spatial replication, you need to 
  
  3. The number of nested spatial replicates.
 
-Below we provide respective three examples for each one of these respective searches.
+Below we provide three examples for each one of these respective cases.
 
 ```{r}
 pplr_browse(tot_spat_rep > 100)
@@ -65,7 +65,9 @@ pplr_browse(n_spat_levs == 3)
 
 #### After downloading data
 
-Users can also explore the spatial and temporal replication of the data more explicitly after downloading it with `pplr_get_data()`. `pplr_site_rep()` provides two options for exploring data that meet temporal replication requirements at a given spatial resolution. You can choose to filter your data by specifying a minimum sampling frequency per year and a minimum number of years that sample with that frequency. Because this function uses the sampling dates to calculate the frequency, it provides additional information that is not contained in the `samplefreq` column of the main metadata table.
[…]
