+++ README.Rmd
@@ -11,7 +11,7 @@ output: github_document
 
 ## Popler
 
-Popler is the R package to browse and query the _Popler_ data base. _Popler_ is a PostgreSQL data base that contains population-level datasets from the US long term ecological research (LTER) network. This package is currently only available on GitHub, but our ultimate goal is to submit it to CRAN.
+`popler` is the R package to browse and query the `popler` data base. `popler` is a PostgreSQL data base that contains population-level datasets from the US long term ecological research (LTER) network. This package is currently only available on GitHub, but our ultimate goal is to submit it to CRAN.
 
 ### Installation
 
@@ -37,9 +37,8 @@ devtools::install_github('AldoCompagnoni/popler')
 
 ------
 
-All exported functions in the `popler` R package use the `pplr_` prefix and lazy and/or tidy evaluation, meaning you do not need to manually quote most inputs. 
-We suggest to start exploring the variables in the data base using `pplr_dictionary()`. This function is mean to explore the metadata variables used to describe the datasets contained in _Popler_. `pplr_dictionary()` works with and without inputs. Without inputs, the function returns a data frame showing a description of each metadata variable. The input of `pplr_dictionary()` is supposed to be the name of one of the metadata variables. When `pplr_dictionary()` is provided with the name of a metadata variable, it returns the possible unique values of the variable.
-Additionally, there is the `pplr_report_dictionary()` function which generates an `Rmd` file and renders it into html. This html contains both the meaning of variables, and their unique values.
+All exported functions in the `popler` R package use the `pplr_` prefix and lazy and/or tidy evaluation, meaning you do not need to manually quote most inputs. We suggest to start exploring the variables in the data base using `pplr_dictionary()`. This function explores the metadata variables used to describe the datasets contained in `popler`. `pplr_dictionary()` works with and without inputs. Without inputs, the function returns a data frame showing a description of each metadata variable. The input of `pplr_dictionary()` is supposed to be the name of one of the metadata variables. When `pplr_dictionary()` is provided with the name of a metadata variable, it returns the possible unique values of the variable.
+Additionally, the `pplr_report_dictionary()` function generates an `Rmd` file and renders it into html. This html contains both the meaning of variables, and their unique values.
 
 ```{r dictionary funs, eval = FALSE}
 
@@ -54,7 +53,7 @@ pplr_dictionary()
 
 ------
 
-Once you are familiar with the meaning and content of _Popler_'s metadata variables, `pplr_browse()` provides the metadata of the studies contained in _Popler_. `pplr_browse()` also works with and without an input. Without input, the function produces a data frame including the metadata variables describing every study currently contained in the _Popler_ database. Note that this data frame is a `tbl` that inherits from the `browse` class. Inputs to `pplr_browse()` allow users to subset this data frame (e.g. `duration_years > 5`). When subsetting, the unique values provided by `pplr_dictionary()` are particularly useful. For more nuanced subsetting of available datasets, the `keyword` argument allows to subset variables using partial matching. Note that `keyword` will act primarily on information contained in the title of studies.
[…]

[…]
