### 2019-04-15 bed5351
Roxygenize grammar updats, further correct mistakes
