+++ man/pop_search.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/util.R
+\name{pop_search}
+\alias{pop_search}
+\title{search}
+\usage{
+pop_search(proj_metadata_key, limit = 10, offset = 0, ...)
+}
+\arguments{
+\item{proj_metadata_key}{project metadata key}
+
+\item{limit}{number of records to return, default: 10}
+
+\item{offset}{record number to start at, default: first record}
+
+\item{...}{curl options passed on to [crul::HttpClient]}
+}
+\description{
+search
+}
+\examples{
+# basic example
+pop_search(proj_metadata_key = 13)
+# pass in curl options for debugging, seeing http request details
+pop_search(proj_metadata_key = 13, verbose = TRUE)
+}
