+++ man/pplr_maps.Rd
@@ -10,11 +10,10 @@ pplr_maps(input, return_plot = FALSE)
 \item{input}{An object created by either \code{pplr_browse()} or
 \code{pplr_get_data()}}
 
-\item{return_plot}{A logical indicating whether to return a copy of the 
-\code{input} data or the \code{ggplot} object created by the function. Use
-\code{TRUE} to return the \code{ggplot} object for subsequent modification.
-Use \code{FALSE} to return an invisible copy of the \code{input} object (
-useful for piping). Default is \code{FALSE}.}
+\item{return_plot}{logical; if \code{TRUE} function returns the \code{ggplot} 
+object for subsequent modification.
+If \code{FALSE}, function returns an invisible copy of the \code{input} 
+object (useful for piping). Default is \code{FALSE}.}
 }
 \value{
 The \code{input} object (invisibly) or a \code{ggplot2} object.
