+++ NAMESPACE
@@ -20,8 +20,7 @@ export(pplr_report_metadata)
 export(pplr_site_rep)
 export(pplr_site_rep_plot)
 export(pplr_summary_table_update)
-importFrom(crul,jsonlite)
-importFrom(crul,tibble)
+importFrom(crul,HttpClient)
 importFrom(dbplyr,translate_sql)
 importFrom(dbplyr,translate_sql_)
 importFrom(dplyr,"%>%")
@@ -55,6 +54,7 @@ importFrom(grid,grid.layout)
 importFrom(grid,grid.newpage)
 importFrom(grid,pushViewport)
 importFrom(grid,viewport)
+importFrom(jsonlite,fromJSON)
 importFrom(lazyeval,lazy_dots)
 importFrom(lubridate,ydm)
 importFrom(magrittr,"%>%")
@@ -71,6 +71,7 @@ importFrom(stats,setNames)
 importFrom(stringr,str_extract_all)
 importFrom(stringr,str_match)
 importFrom(stringr,str_split)
+importFrom(tibble,as_tibble)
 importFrom(tidyr,nest)
 importFrom(utils,View)
 importFrom(utils,bibentry)
