### 2019-05-05 #254 @AldoCompagnoni
Dear @noamross ,
Thanks for checking in! We are done with points 1-4, but we are currently de-bugging issues with the database itself. A bit after the review, the database broke due to hacking/lack of space. We just set a new instance of the database with increased security settings, but we are still working to set up the API; at the same time, we are planning to upgrade to a AWS instance with twice as much RAM and disk space in the upcoming 2 months. This upgrade is needed to prevent new crashes due to disk space, and increase the speed of downloads.

We decided not to address point 5 for a reason that we consider very important: we realized that addressing it would interphere with the way we designed the package to work. Writing an expression such as `pplr_browse( duration_years > 20 )` does not just subset a data frame, but produces an S3 class with attributes that can be used by functions `pplr_get_data()`, `pplr_metadata_url()`, and `pplr_citation()`. We present this workflow in Figure 2 of popler's manuscript draft (https://github.com/texmiller/popler-ms/blob/master/popler_ms.pdf).

I have a writing retreat next week, so I expect to get back by May 19th at the latest.
