### 2019-05-05 3514905
Update imports for API functions
