### 2019-05-14 f433005
Restructure readme file around 'finding, retrieve,and use' data. Link to manuscript draft and vignettes
