+++ README.md
@@ -3,9 +3,11 @@
 [![Build Status](https://travis-ci.org/AldoCompagnoni/popler.svg?branch=master)](https://travis-ci.org/AldoCompagnoni/popler) [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/AldoCompagnoni/popler?branch=master&svg=true)](https://ci.appveyor.com/project/AldoCompagnoni/popler) [![Coverage status](https://codecov.io/gh/AldoCompagnoni/popler/branch/master/graph/badge.svg)](https://codecov.io/github/AldoCompagnoni/popler?branch=master)
 
 Popler
-------
+======
 
-Popler is the R package to browse and query the *Popler* data base. *Popler* is a PostgreSQL data base that contains population-level datasets from the US long term ecological research (LTER) network. This package is currently only available on GitHub, but our ultimate goal is to submit it to CRAN.
+`popler` is the R package to browse and query the `popler` data base. `popler` is a PostgreSQL data base that contains population-level datasets from the US long term ecological research (LTER) network. This package is currently only available on GitHub, but our ultimate goal is to submit it to CRAN. A more detailed explanation on the structure of the `popler` online database is contained in the [dratf of the manuscript presenting popler](https://github.com/texmiller/popler-ms/blob/master/popler_ms.pdf) and in the [dedicated vignette](https://github.com/AldoCompagnoni/popler/blob/master/vignettes/popler-database-structure.Rmd).
+
+The package `popler` aims to facilitate finding, retrieving, and using time-series data of population abundance associated with the US LTER network. To *find datasets*, the functions in `popler` aid in understanding and browsing the metadata information referring to each dataset. To *retrieve* data, the function `pplr_get_data()` downloads single datasets or groups of datasets. These downloads share the same data structure. To *use* downloaded data, the package provides ancillary functions to consult and cite the original data sources, examine the temporal replication of each data set, and methods for a couple of `dplyr` verbs to assist with data manipulation.
 
 ### Installation
 
@@ -24,11 +26,14 @@ if(!require(devtools, quietly = TRUE)) {
 devtools::install_github('AldoCompagnoni/popler')
 ```
 
-### Dictionary of variables
+Finding datasets
+================
 
 ------------------------------------------------------------------------
 
-All exported functions in the `popler` R package use the `pplr_` prefix and lazy and/or tidy evaluation, meaning you do not need to manually quote most inputs. We suggest to start exploring the variables in the data base using `pplr_dictionary()`. This function is mean to explore the metadata variables used to describe the datasets contained in *Popler*. `pplr_dictionary()` works with and without inputs. Without inputs, the function returns a data frame showing a description of each metadata variable. The input of `pplr_dictionary()` is supposed to be the name of one of the metadata variables. When `pplr_dictionary()` is provided with the name of a metadata variable, it returns the possible unique values of the variable. Additionally, there is the `pplr_report_dictionary()` function which generates an .Rmd file and renders it into html. This html contains both the meaning of variables, and their unique values.
+##### Dictionary of variables
+
[…]
