### 2019-05-15 c9ffece
Sub '%' with 'percent' in doi_citation entry
