### 2019-05-15 e0dd0c6
Add dplyr::bind_rows import in NAMESPACE
