+++ NAMESPACE
@@ -25,6 +25,7 @@ importFrom(dbplyr,translate_sql)
 importFrom(dbplyr,translate_sql_)
 importFrom(dplyr,"%>%")
 importFrom(dplyr,bind_cols)
+importFrom(dplyr,bind_rows)
 importFrom(dplyr,filter)
 importFrom(dplyr,group_by)
 importFrom(dplyr,mutate)
