+++ .travis.yml
@@ -29,7 +29,7 @@ matrix:
     r: oldrel   
   include:
     os: osx
-    osx_image: xcode8
+    osx_image: xcode8.3
     brew_packages: libgit2
     
 notifications:
