### 2019-05-15 b9caacd
Make lat_lter/lng_lter numeric; replace 'F' with'FALSE'
