+++ DESCRIPTION
@@ -41,7 +41,8 @@ Suggests:
     mapproj (>= 1.2.6),
     testthat(>= 2.0.0),
     covr (>= 3.1.0),
-    stubthat (>= 1.2.1)
+    stubthat (>= 1.2.1),
+    mockr (>= 0.1)
 LazyData: true
 RoxygenNote: 6.1.1
 VignetteBuilder: knitr
