### 2019-05-17 d2c18b6
Correct format issues on sum_tbl associated with API
