### 2019-05-17 44fd940
Render DB vignette without readxl package
