### 2019-05-17 3ee26c5
Remove chunk of code to install readxl
