### 2019-05-17 6f395bb
Declare use of txtProgressBar setTxtProgressBar
