### 2019-05-17 7bb53f8
Solve global variable notes; progress bar imports
