### 2019-05-17 cfaaa75
Call pplr_query using :::
