+++ tests/testthat/test-reporting.R
@@ -6,7 +6,7 @@ library(ggplot2)
 # "mock" download of data object using a stub ---------------------------
 
 # Invoke the stub function with the function that needs to be mocked
-pplr_query_stub <- stubthat::stub( popler::pplr_query )
+pplr_query_stub <- stubthat::stub( popler:::pplr_query )
 
 # Define the behavior of the stub
 pplr_query_stub$withArgs( proj_id = 1 )$returns( int.data$data_1 )
