### 2019-05-17 919555b
Remove RPostgreSQL, now useless import
