### 2019-05-17 fd3fe66
Add import for text progress bar
