+++ NAMESPACE
@@ -77,5 +77,8 @@ importFrom(tidyr,nest)
 importFrom(utils,View)
 importFrom(utils,bibentry)
 importFrom(utils,browseURL)
+importFrom(utils,globalVariables)
 importFrom(utils,person)
+importFrom(utils,setTxtProgressBar)
 importFrom(utils,toBibtex)
+importFrom(utils,txtProgressBar)
