### 2019-05-17 c6dc072
Remove pryr:colwise
