+++ DESCRIPTION
@@ -13,7 +13,7 @@ URL: https://github.com/AldoCompagnoni/popler
 BugReports: https://github.com/AldoCompagnoni/popler/issues
 Depends: R (>= 3.0.0)
 Imports:
-    crul (>= 0.5.2),
+    crul (>= 0.7.4),
     dbplyr (>= 1.1.0), 
     dplyr (>= 0.5.0), 
     ggplot2 (>= 3.0.0),
