### 2019-05-19 a58da80
Add description on database structure; picture of pplr_site_rep_plot()
