+++ README.Rmd
@@ -12,7 +12,22 @@ output: github_document
 # Popler
 
 `popler` is the R package to browse and query the `popler` data base. `popler` is a PostgreSQL data base that contains population-level datasets from the US long term ecological research (LTER) network. This package is currently only available on GitHub, but our ultimate goal is to submit it to CRAN.
-A more detailed explanation on the structure of the `popler` online database is contained in the [dratf of the manuscript presenting popler ](https://github.com/texmiller/popler-ms/blob/master/popler_ms.pdf) and in the [dedicated vignette](https://github.com/AldoCompagnoni/popler/blob/master/vignettes/popler-database-structure.Rmd).
+A detailed explanation on the structure of the `popler` online database is contained in the [dratf of the manuscript presenting popler ](https://github.com/texmiller/popler-ms/blob/master/popler_ms.pdf) and in the [dedicated vignette](https://github.com/AldoCompagnoni/popler/blob/master/vignettes/popler-database-structure.Rmd). 
+The `popler` database is organized around four types of tables:
+
+A. The tables containing information on population abundance. Population abundance can be of five types: count, biomass, cover, density, and at the individual-level.
+
+B. A table containg to link each abundance value to the taxonomic units it refers to. In `popler`, "taxonomic unit" generally refers to a species.
+
+C. The location of each "site". With "site" we refers to the largest spatial replicate available for a dataset. Many datasets provide abundance data from more than one site.
+
+D. Metadata information referring to each separate dataset.
+
+```{r, echo=FALSE}
+knitr::include_graphics("vignettes/img/schema.png")
+```
+
+# Popler rationale
 
 The package `popler` aims to facilitate finding, retrieving, and using time-series data of population abundance associated with the US LTER network.
 To _find datasets_, the functions in `popler` aid in understanding and browsing the metadata information referring to each dataset.
@@ -39,22 +54,44 @@ devtools::install_github('AldoCompagnoni/popler')
 
 ```
 
+All exported functions in the `popler` R package use the `pplr_` prefix. Moreover, the functions use lazy and/or tidy evaluation, meaning you do not need to manually quote most inputs.
+
+
 # Finding datasets
 
 ------
 
 ##### Dictionary of variables
 
-All exported functions in the `popler` R package use the `pplr_` prefix and lazy and/or tidy evaluation, meaning you do not need to manually quote most inputs. We suggest to start exploring the variables in the data base using `pplr_dictionary()`. This function explores the metadata variables used to describe the datasets contained in `popler`. `pplr_dictionary()` works with and without inputs. Without inputs, the function returns a data frame showing a description of each metadata variable. The input of `pplr_dictionary()` is supposed to be the name of one of the metadata variables. When `pplr_dictionary()` is provided with the name of a metadata variable, it returns the possible unique values of the variable.
-Additionally, the `pplr_report_dictionary()` function generates an `Rmd` file and renders it into html. This html contains both the meaning of variables, and their unique values.
+We suggest to start exploring the metadata variables describing each dataset in popler using `pplr_dictionary()`. This function works in two ways:
 
-```{r dictionary funs, eval = FALSE}
+1. It provides a general description of each metadata variable. This happens when this function is called without arguments; for example, when calling `pplr_dictionary()`.
+1. It provides the possible values of its unquoted arguments. For example, when calling `pplr_dictionary( proj_metadata_key )`.
+
+The output of `pplr_dictionary()` is a data frame showing a description of each metadata variable. 
+
+
+```{r load popler, eval = TRUE, include = FALSE}
+library(popler)
+```
+
+```{r dictionary funs, eval = TRUE}
 
 library(popler)
-
 pplr_dictionary()
 
 ```
+
[…]
