### 2019-05-19 849eb43
Add curl 3.3 'suggest': try to pass AppVeyor
