### 2019-05-19 2b8c328
Depends on R 3.5.3 (for AppVeyor)
