+++ DESCRIPTION
@@ -11,7 +11,7 @@ Authors@R: c(
 Description: Browse and query the popler database.
 URL: https://github.com/AldoCompagnoni/popler
 BugReports: https://github.com/AldoCompagnoni/popler/issues
-Depends: R (>= 3.6.0)
+Depends: R (>= 3.5.3)
 Imports:
     crul (>= 0.7.4),
     dbplyr (>= 1.1.0), 
