### 2019-05-19 #254 @AldoCompagnoni
I am checking in as announced, but regretfully we have found a substantial issue with the API. We think it is more practical to provide a rundown of the changes we've made to the package once the package itself is running perfectly.

We'll be back to this as soon as possible, hopefully within mid week.
