### 2019-05-20 55172fa
Add curl import (for AppVeryor)
