+++ DESCRIPTION
@@ -14,6 +14,7 @@ BugReports: https://github.com/AldoCompagnoni/popler/issues
 Depends: R (>= 3.5.3)
 Imports:
     crul (>= 0.7.4),
+    curl (>= 3.3),
     dbplyr (>= 1.1.0), 
     dplyr (>= 0.5.0), 
     ggplot2 (>= 3.0.0),
@@ -40,8 +41,7 @@ Suggests:
     testthat(>= 2.0.0),
     covr (>= 3.1.0),
     stubthat (>= 1.2.1),
-    mockr (>= 0.1),
-    curl (>= 3.3)
+    mockr (>= 0.1)
 LazyData: true
 RoxygenNote: 6.1.1
 VignetteBuilder: knitr
