### 2019-06-02 e85117f
Fix download by 1.using integer indices;2.allow download data with rows that are multiples of 1000
