### 2019-06-02 5a3368a
Roxygenize two new functions starting with pplr_
