+++ man/pplr_summary.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/util.R
-\name{pop_summary}
-\alias{pop_summary}
+\name{pplr_summary}
+\alias{pplr_summary}
 \title{search}
 \usage{
-pop_summary(limit = 10, offset = 0, ...)
+pplr_summary(limit = 10, offset = 0, ...)
 }
 \arguments{
 \item{limit}{number of records to return, default: 10}
@@ -18,7 +18,7 @@ search
 }
 \examples{
 # basic example
-pop_summary()
+pplr_summary()
 # pass in curl options for debugging, seeing http request details
-pop_summary(verbose = TRUE)
+pplr_summary(verbose = TRUE)
 }
