+++ man/pplr_search.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/util.R
-\name{pop_search}
-\alias{pop_search}
+\name{pplr_search}
+\alias{pplr_search}
 \title{search}
 \usage{
-pop_search(proj_metadata_key, limit = 10, offset = 0, ...)
+pplr_search(proj_metadata_key, limit = 10, offset = 0, ...)
 }
 \arguments{
 \item{proj_metadata_key}{project metadata key}
@@ -20,7 +20,7 @@ search
 }
 \examples{
 # basic example
-pop_search(proj_metadata_key = 13)
+pplr_search(proj_metadata_key = 13)
 # pass in curl options for debugging, seeing http request details
-pop_search(proj_metadata_key = 13, verbose = TRUE)
+pplr_search(proj_metadata_key = 13, verbose = TRUE)
 }
