+++ NAMESPACE
@@ -6,8 +6,6 @@ S3method(mutate,browse)
 S3method(mutate,get_data)
 export(filter)
 export(mutate)
-export(pop_search)
-export(pop_summary)
 export(pplr_browse)
 export(pplr_citation)
 export(pplr_cov_unpack)
@@ -17,8 +15,10 @@ export(pplr_maps)
 export(pplr_metadata_url)
 export(pplr_report_dictionary)
 export(pplr_report_metadata)
+export(pplr_search)
 export(pplr_site_rep)
 export(pplr_site_rep_plot)
+export(pplr_summary)
 export(pplr_summary_table_update)
 importFrom(crul,HttpClient)
 importFrom(dbplyr,translate_sql)
