### 2019-06-02 8a8de89
Specify 'PKGTYPE: source' in oldrel
