+++ appveyor.yml
@@ -21,6 +21,7 @@ environment:
 
     - R_VERSION: release
       R_ARCH: i386
+      
 
 # R-devel failing for windows builds right now. appears to be a timeout issue,
 # really not sure why this is. It doesn't appear to be related to the package
@@ -34,9 +35,11 @@ environment:
 
     - R_VERSION: oldrel
       R_ARCH: x64
+      PKGTYPE: source
 
     - R_VERSION: oldrel
       R_ARCH: i386
+      PKGTYPE: source
 
 build_script:
   - travis-tool.sh install_deps
