### 2019-06-10 513ee8e
Remove curl from imports
