+++ DESCRIPTION
@@ -11,10 +11,9 @@ Authors@R: c(
 Description: Browse and query the popler database.
 URL: https://github.com/AldoCompagnoni/popler
 BugReports: https://github.com/AldoCompagnoni/popler/issues
-Depends: R (>= 3.5.3)
+Depends: R (>= 3.5.0)
 Imports:
     crul (>= 0.7.4),
-    curl (>= 3.3),
     dbplyr (>= 1.1.0), 
     dplyr (>= 0.5.0), 
     ggplot2 (>= 3.0.0),
