### 2019-06-10 b295dff
Remove purrr:map2
