### 2019-06-10 b4c64c1
Remove dependency on lubridate
