+++ NAMESPACE
@@ -57,7 +57,6 @@ importFrom(grid,pushViewport)
 importFrom(grid,viewport)
 importFrom(jsonlite,fromJSON)
 importFrom(lazyeval,lazy_dots)
-importFrom(lubridate,ydm)
 importFrom(magrittr,"%>%")
 importFrom(purrr,pmap_chr)
 importFrom(rlang,"!!!")
