### 2019-06-10 e5b1170
Remove dependency on lubridate
