+++ DESCRIPTION
@@ -20,7 +20,6 @@ Imports:
     grid (>= 3.3.0),
     jsonlite (>= 1.5),
     lazyeval (>= 0.2.0),
-    lubridate (>= 1.7.4),
     magrittr (>= 1.5.0),
     purrr (>= 0.2.5),
     rlang (>= 0.2.1),
