### 2019-06-10 #254 @AldoCompagnoni
Dear @cgries, @bpbond, and @noamross,
Below I illustrate our revision of the code for the popler R package. I will split this report point by point, as opposted to the 4 points in which I had originally summarized the reviews. In order:

**goodpractice::gp()**
These checks show only issues with the tests - but our codecov shows 72% of our code is covered by tests, which @noamross suggested is fine.

**devtools::test()**
All 99 tests pass without warnings/notes on my machine.
Following the wonderful suggestion by @bpbond, we have implemented all tests that rely on a download using STUBS. Tests now run in about 140 seconds.

**devtools::check()**
Checks pass, but we get this NOTE:

Non-standard file/directory found at top level:
  'README_files'

because we use a directory, README_files, to store one of the two figures that we added to the README. We are unaware if this is a substantial issue, so please let us know!

*Documentation: README and VIGNETTES*

All reviewers, @cgries, @bpbond, and @noamross, expressed concerns about our documentation. Therefore:
1. We included a new vignette dedicated to the structure of the popler database, we included a succinct representation of the schema in the README, and we pointed to the vignettes and manuscript draft whenever logical. We also added an additional figure to the README to show the output of function `pplr_site_rep_plot()`.
2. We reviewed all the documentation to improve clarity, catch grammatical mistakes, and typos.

**Issues with the downloads**

The issues originated from:
1. First, and foremost, our AWS instance initially did not have enough disk space: only 8GB. We increased disk space to 20GB, and we are now using only 44% of the disk space available, which increased the speed of downloads. From the examples mentioned above, `Antarctica <- pplr_get_data(lterid == 'PAL')` takes less than 2 minutes, `poa_data <- pplr_get_data(poa_metadata)` takes 2 minutes. Note that `BNZ <- pplr_get_data(lterid == 'BNZ')` and `SEV_studies <- pplr_get_data(lterid == 'SEV')` take 6 and 10. We consider these download times to be tolerable, because the size of these datasets is 70MB and 120MB, respectively.
2. The downloader contained in popler had two mundane mistakes associated with downloads of large datasets, and one dataset whose number of rows was an exact multiple of 1000. This is my bad, as I had not carefully tested the new downloader associated with the API on each single dataset. Rather, I had relied solely on the unit tests.

**Report capability**

We removed the warnings found by @bpbond when running `pplr_browse(report=TRUE)`. This was caused by an issue with the encoding of a single study title, and on the standard warnings on multiple URLS produced by function `pplr_metadata_url()`.
Regading the warnings on multiple URLs, we have retained these when using pplr_metadata_url() with objects produced by both `pplr_browse()` and `pplr_get_data()`. We think these warnings are informative (e.g., execute `pplr_browse(proj_metadata_key == 297 ) %>% pplr_metadata_url()` ).

**Dependencies**

Thank you @bpbond for commenting on dependencies: we did find that plyr and lubridate were not needed.

**Duplication of functionality**

Regarding substituting
`pplr_browse(duration_years > 20)`
with
`pplr_browse() %>% filter(duration_years > 20)`

We decided not to implement this because it would remove one of the characteristics that we purpusefully built into popler: `pplr_browse(duration_years > 20)` provides an object of class `popler` that contains, as one of its attributes, the logical statement of the query (in this specific case `duration_years > 20`). This attribute is then used inside of `pplr_get_data()`.

**Curious observation about dataset number 133**

This is a good catch: we've noticed that in 14 datasets the authors/authors_contact in our project metadata, and the author reported in the citation do not correspond.
This occurs whenever the original data does not report the email of the first author of the dataset. In that case, we reported as "author", the person for which we had an email contact (usually one of the co-authors).
We think this is Ok, but @cgries please let us know if there are reasons for this to change: the alternative it to list the author, but have an `NA` in the authors_contact column.
