### 2019-06-11 e0f8f4e
Add Rtools to Appveyor builds
