+++ appveyor.yml
@@ -15,6 +15,7 @@ cache:
 
 # Adapt as necessary starting from here
 environment:
+  USE_RTOOLS: true
   matrix:
     - R_VERSION: release
       R_ARCH: x64
