### 2019-06-25 #254 @noamross
Thanks for the update, @AldoCompagnoni! @cgries and @bpbond, can you look at the changes and tell us if they address your previous comments?
