### 2019-06-27 #254 @bpbond
The changes laid out by @AldoCompagnoni above sound like they'll fully address my comment, yes.
