### 2019-06-27 #254 @cgries
I agree, the changes look like they are addressing my concerns. To answer the last question, I would resort to calling the person you have an e-mail address for the contact rather than author. But that is so minor, that I don't feel it needs any further addressing.
