### 2019-06-28 #254 @noamross
Thank you @bpbond and @cgries.

@AldoCompagnoni, I am trying to do final editor checks, but in building the vignettes, and also any `pplr_get_data()` call, I'm getting `Error: Bad Request (HTTP 400)`.  Is the service down?
