### 2019-06-28 #254 @AldoCompagnoni
@noamross, @bpbond and @cgries,
Thanks for getting back!

@noamross: yes, the service is down, but this time it is the API and not the database (I bet it is something with the web server). I am traveling right now, but I'll fix this issue tomorrow, and I'll try to find a way to get notified when the API stops working.

Apologies for the inconvenience, but this is part of the process (I'm pretty sure I downloaded data from popler just two mondays ago, so this came out of the blue).
