### 2019-07-03 #254 @AldoCompagnoni
Thank you @noamross,
And enjoy your vacation!

@annakrystalli, we have resolved the issue reported above by Noam: as expected, the web server was down. We are still evaluating options to automatically check when the API stops working (suggestions are welcome).

Note that I am on vacation until July 13th, so I might be slow in replying up to that date.
