### 2019-07-04 #254 @annakrystalli
Actually, one last thing @AldoCompagnoni. I would expect, given the development during the review process that the package version would have changed since submission to reflect that. For more details on how to track changes in your package through versioning and `NEWS` see the [relevant section](https://devguide.ropensci.org/releasing.html#versioning) in the dev guide.
