### 2019-07-04 #254 @stefaniebutland
Hi @AldoCompagnoni. I'm rOpenSci's Community Manager, following up on blog post invitation. Noam Ross suggested a post "talking about managing both the service and the client package, but also the LTER network and such" would be well-received.

If you're interested, right now I have Tues July 23 available for publication. Ideally with a draft for me to review submitted by July 16. Later dates are ok too.

Let me know what you think. Happy to answer any questions.
