### 2019-07-23 6430707
Update NEWS with ropensci review changes
