+++ NEWS.md
@@ -1,4 +1,31 @@
-## Popler 0.1.0
+Popler 0.2.0 (2019-07-23)
+=========================
 
-- Introduces the first stable release of the `popler` _R_ package for querying
-the PostgreSQL data base of the same name.
+### NEW FEATURES
+
+  * This new version of `Popler` queries the online database via an API
+
+### MINOR IMPROVEMENTS
+
+  * Faster tests through the use of stubs
+
+### BUG FIXES
+
+  * Fix bugs contained in `report_metadata`
+
+### DEPRECATED AND DEFUNCT
+
+  * `hello_world()` now deprecated and will be removed in a
+     future version, use `hello_mars()`
+
+### DOCUMENTATION FIXES
+
+  * Created new vignette, `popler-database-structure`, to provide a full documentation the database structure
+  * Updated README file to include information on database structure
+  * Removed typos and clarified functions documentation
+
+
+Popler 0.1.0 
+=========================
+
+  * Introduces the first stable release of the `popler` _R_ package for querying the PostgreSQL data base of the same name.
