### 2019-07-23 cadbf72
Make TravisCI points to ropensci repository
