### 2019-07-23 7a28a52
Add Benjamin Bond-Lamberty as reviewer; change URL to github.com/ropensci/popler
