+++ DESCRIPTION
@@ -7,10 +7,13 @@ Authors@R: c(
   person("Bibian", "Andrew", email = "ajbibian@rice.edu", role= c("aut")),
   person("Ochocki", "Brad", email = "brad.ochocki@rice.edu", role= c("aut")),
   person("Levin", "Sam", email = "sam.levin@idiv.de", role = c("aut")),
-  person("Miller", "Tom", email = "tom.miller@rice.edu", role= c("aut")))
+  person("Miller", "Tom", email = "tom.miller@rice.edu", role= c("aut")),
+  person("Benjamin", "Bond-Lamberty", role = "rev",
+          comment = "Ben reviewed the package for ropensci, see <https://github.com/ropensci/software-review/issues/254>")
+          )
 Description: Browse and query the popler database.
-URL: https://github.com/AldoCompagnoni/popler
-BugReports: https://github.com/AldoCompagnoni/popler/issues
+URL: https://github.com/ropensci/popler
+BugReports: https://github.com/ropensci/popler/issues
 Depends: R (>= 3.5.0)
 Imports:
     crul (>= 0.7.4),
