### 2019-07-24 4ebc713
Update travisCI badge
