+++ README.Rmd
@@ -50,7 +50,7 @@ if(!require(devtools, quietly = TRUE)) {
   install.packages(devtools)
 }
 
-devtools::install_github('AldoCompagnoni/popler')
+devtools::install_github('ropensci/popler')
 
 
 ```
