### 2019-07-24 #254 @AldoCompagnoni
Dear @annakrystalli,
Thank you for your helpful last comments, which I believe we have addressed in full!

@stefaniebutland, we would be happy to submit a blog post for rOpenSci, but as you see, I am well over the July 16 deadline! What other dates do you have available? I'll be at Ecological Society of America 2019 (August 11st-19th), so a deadline before or after the ESA meeting would be optimal. Thank you in advance!
