### 2019-07-26 5b61748
Add Corinna Gries as a reviewer of the package
