### 2019-07-26 #254 @AldoCompagnoni
@annakrystalli,
I hope that the ticks below, quoting your approval comment, address your request as hoped!

Moreover, I have credited both reviewers, and I have also updated the `NEWS.md` file. However, in this last case I have not yet produced an release as suggested by the [relevant section](https://devguide.ropensci.org/releasing.html#versioning) of the development guide. I would like to associate that with a CRAN release (the package has not been pushed to CRAN yet).

Let me know if I have left out relevant actions!

>
>
> Approved! raised_hands Thanks @AldoCompagnoni for submitting and @bpbond & @cgries for your reviews!
>
> To-dos:
>
>     * [ ]  Transfer the repo to rOpenSci's "ropensci" GitHub organization under "Settings" in your repo.  I have invited you to a team that should allow you to do so.  You'll be made admin once you do.
>
>     * [ x ]  Add a rOpenSci review badge to your README:
>       ```
>       ```
>
>     * [ x ]  Add the rOpenSci footer to the bottom of your README
>       ```
>       [![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
>       ```
>
>     * [ x ]  Fix any links in badges for CI and coverage to point to the ropensci URL. We no longer transfer Appveyor projects to ropensci Appveyor account so after transfer of your repo to rOpenSci's "ropensci" GitHub organization the badge should be:
>       ```
>       [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/popler?branch=master&svg=true)](https://ci.appveyor.com/project/AldoCompagnoni/popler)
>       ```
>
>     * [ x ]  We're starting to roll out software metadata files to all ropensci packages via the Codemeta initiative, see https://github.com/ropensci/codemetar/#codemetar for how to include it in your package, after installing the package - should be easy as running codemetar::write_codemeta() in the root of your package.
>
>
> Should you want to acknowledge your reviewers in your package DESCRIPTION, you can do so by making them `"rev"`-type contributors in the `Authors@R` field (with their consent). More info on this [here](https://ropensci.github.io/dev_guide/building.html#authorship).
>
> Welcome aboard! We'd love to host a blog post about your package - either a [short introduction to it with one example](https://ropensci.org/tech-notes/) or [a longer post with some narrative about its development or something you learned, and an example of its use](https://ropensci.org/blog/). If you are interested, review [the instructions](https://github.com/ropensci/roweb2#contributing-a-blog-post), and tag @stefaniebutland in your reply. She will get in touch about timing and can answer any questions.
>
> We've started putting together a gitbook with our best practice and tips, [this chapter](https://ropensci.github.io/dev_guide/collaboration.html) starts the 3d section that's about guidance for after onboarding. Please tell us what could be improved, the corresponding repo is [here](https://github.com/ropensci/dev_guide).
>
> Any questions, just let me know. And enjoy your holidays! beach_umbrella
