### 2019-07-30 #254 @stefaniebutland
Hi @AldoCompagnoni. Would you be able to submit a draft blog post via pull request before ESA? I may have a publication slot open for Aug 6 or 13th and prefer to receive a draft one week prior to publication so I can review.

Please suggest a date for draft submission and we can proceed from there. Thank you!
