### 2019-08-02 #254 @AldoCompagnoni
@stefaniebutland,
A draft on August 6th to be published on August 13th would work for me! Expect my draft, and please let me know if there is additional information I should be aware of!
