### 2019-08-02 #254 @stefaniebutland
Sounds good @AldoCompagnoni. I've marked my calendar to look for draft on Aug 6.

Content suggestion from Noam Ross: a post "talking about managing both the service and the client package, but also the LTER network and such" would be well-received. It's important not to repeat information that's in the vignette; refer to it where needed.

Editorial and technical information about preparing your submission: https://github.com/ropensci/roweb2#contributing-a-blog-post

Don't hesitate to ask me questions here or in Slack. I'll be away Mon Aug 5.

Thank you for doing this extra work. I hope the post will get more eyes on your work.
