### 2019-09-30 f32e07a
delete accidental insertions
