+++ DESCRIPTION
@@ -14,7 +14,7 @@ Authors@R: c(
           comment = "Corinna reviewed the package for ropensci, see <https://github.com/ropensci/software-review/issues/254>")
           )
 Description: Browse and query the popler database.
-URL: https://github.com/ropensci/popler
+URL: https://github.com/ropensci/popler, https://docs.ropensci.org/popler/
 BugReports: https://github.com/ropensci/popler/issues
 Depends: R (>= 3.5.0)
 Imports:
