### 2019-10-15 ac1345e
small documentation fix
